Fix: segfault and memory leaks

+++ includes/ftrace.h
@@ -629,7 +629,7 @@ void print_map(link_t *map_list);
 void delete_all_map(link_t *map_list);
 void free_map(void *obj);
 
-void handle_opcode(regs_t regs, uint64_t rip, pid_t pid, link_t **stack);
+void handle_opcode(regs_t regs, uint16_t rip, pid_t pid, link_t **stack);
 void handle_signal(pid_t pid);
 char *get_function_name(const char *bin_name, uint64_t val);
